Create authentication_console.js
Using node.js platform for the console application. The reason why we choose node.js is that it’s
single threaded and non blocking. This make request response flow smoother and faster and it
can be used to process a large number of requests from IOT device.Another reason is that it’s
event driven allowing us to run various task without waiting for other tasks to complete .